sshd monitoring: remove preauth logs
The original sshd monitoring configuration (monit 5.2.5) monitored
sshd by attempting to log into localhost port 22 seen in log files
like this:

  Feb  5 00:55:45 cumulus sshd[24328]: Connection closed by 127.0.0.1
  [preauth]

This pollutes the logs and is generally an unnecessary check when
verifying the pid is available. Thus change the check to validate the
pid and if not valid restart the daemon.

Signed-off-by: Jonathan Toppins <jtoppins@cumulusnetworks.com>