sysloadavg: fix SEGFAULT caused by incorrect printf type
Commit 12e10224738fcf43919982ee3207d65ed18e2667 added support for
a new resource "sysloadavg". In http/cervlet.c line 1843 the call
to out_print in the format specifier defined a float or double ("%.1f")
be provided in the 4th argument. The original value provided was of type
int.

The fix is to explicitly typecast the 4th argument to type double.

Closes: https://tickets.cumulusnetworks.com/browse/CM-5345
Fixes: 12e10224738fcf43919982ee3207d65ed18e2667
Signed-off-by: Jonathan Toppins <jtoppins@cumulusnetworks.com>