jdoorc: only load config files that have a specific suffix
The example configuration file was loading every file in the
/etc/jdoo/jdoorc.d/ directory. This can lead to errors being reported by
jdoo if a package being upgraded is still in the unpacked state (for
Debian) and jdoo is restarted or configuration reloaded. The result will
be something like:
  /etc/jdoo/jdoorc.d/foo.rc:1: Error: service name conflict, foo
    already defined '/var/run/foo.pid'

Since some package managers, dpkg in this case, use files like
foo.rc-dpkg-new to stage new files before upgrading the configuration
file we need jdoo to ignore these temporary files. The easiest way to do
this is to force a well known suffix to be added to config files.

Signed-off-by: Jonathan Toppins <jtoppins@cumulusnetworks.com>